+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.7.1 Release (08/xx/2026)
 - Added new option to extract todos from notes that can be turned into Google Tasks
-- Fix for note-to-note linking format for Apple Notes
+- Added warnings that extracting hashtags to labels (-h) and creating Tasks notes (-ct) will alter your Keep notes
+- Fix for note-to-note linking format (>>) for Apple Notes
 - Fixed spelling errors in prompts
 
 ## 0.7.0 Release (05/23/2026)